+++ inc/Estruturas.hpp
@@ -4,12 +4,12 @@
 
 typedef struct Aresta {
     int destino;
-    int peso;
+    double peso;
 }Aresta;
 
 typedef struct Distancia {
     int origem; 
     int destino; 
-    int valor = -1; 
+    double valor = -1; 
 }Distancia; 
 #endif
